requirements.txt
discord.py
langchain
openai
chromadb
python-dotenv
